Enable light weight rules engine for the Smart Home sensors rules
Rules engine takes rules written in JSON format as input and evaluate
payload data received from the sensors with each rule. When the rule
conditions are met, it triggers an action registered for that rule.
